new: refactor mocks
