+++ mocks/constraint.go
@@ -1,0 +1,9 @@
+package mocks
+
+import (
+	"net/http"
+)
+
+type verifier interface {
+	Verify(r *http.Request) []error
+}
