+++ mocks/constraint_nop.go
@@ -1,0 +1,11 @@
+package mocks
+
+import (
+	"net/http"
+)
+
+type nopConstraint struct{}
+
+func (c *nopConstraint) Verify(r *http.Request) []error {
+	return nil
+}
