+++ mocks/strategy.go
@@ -1,0 +1,14 @@
+package mocks
+
+import (
+	"net/http"
+)
+
+type ReplyStrategy interface {
+	HandleRequest(w http.ResponseWriter, r *http.Request) []error
+}
+
+type contextAwareStrategy interface {
+	ResetRunningContext()
+	EndRunningContext() []error
+}
