+++ mocks/strategy_template_reply.go
@@ -11,6 +11,21 @@ import (
 	"github.com/Masterminds/sprig/v3"
 )
 
+func NewTemplateReply(content string, statusCode int, headers map[string]string) (ReplyStrategy, error) {
+	tmpl, err := template.New("").Funcs(sprig.GenericFuncMap()).Parse(content)
+	if err != nil {
+		return nil, fmt.Errorf("template syntax error: %w", err)
+	}
+
+	strategy := &templateReply{
+		replyBodyTemplate: tmpl,
+		statusCode:        statusCode,
+		headers:           headers,
+	}
+
+	return strategy, nil
+}
+
 type templateReply struct {
 	replyBodyTemplate *template.Template
 	statusCode        int
@@ -39,21 +54,6 @@ func (tr *templateRequest) Json() (map[string]interface{}, error) {
 	}
 
 	return tr.jsonData, nil
-}
-
-func newTemplateReply(content string, statusCode int, headers map[string]string) (ReplyStrategy, error) {
-	tmpl, err := template.New("").Funcs(sprig.GenericFuncMap()).Parse(content)
-	if err != nil {
-		return nil, fmt.Errorf("template syntax error: %w", err)
-	}
-
-	strategy := &templateReply{
-		replyBodyTemplate: tmpl,
-		statusCode:        statusCode,
-		headers:           headers,
-	}
-
-	return strategy, nil
 }
 
 func (s *templateReply) executeResponseTemplate(r *http.Request) (string, error) {
