+++ mocks/strategy_uri_vary_reply.go
@@ -1,0 +1,44 @@
+package mocks
+
+import (
+	"net/http"
+	"strings"
+)
+
+func NewUriVaryReplyStrategy(basePath string, variants map[string]*Definition) ReplyStrategy {
+	return &uriVaryReply{
+		basePath: strings.TrimRight(basePath, "/") + "/",
+		variants: variants,
+	}
+}
+
+var _ contextAwareStrategy = (*uriVaryReply)(nil)
+
+type uriVaryReply struct {
+	basePath string
+	variants map[string]*Definition
+}
+
+func (s *uriVaryReply) HandleRequest(w http.ResponseWriter, r *http.Request) []error {
+	for uri, def := range s.variants {
+		uri = strings.TrimLeft(uri, "/")
+		if s.basePath+uri == r.URL.Path {
+			return def.Execute(w, r)
+		}
+	}
+	return unhandledRequestError(r)
+}
+
+func (s *uriVaryReply) ResetRunningContext() {
+	for _, def := range s.variants {
+		def.ResetRunningContext()
+	}
+}
+
+func (s *uriVaryReply) EndRunningContext() []error {
+	var errs []error
+	for _, def := range s.variants {
+		errs = append(errs, def.EndRunningContext()...)
+	}
+	return errs
+}
